Import of MIRK/Alvis-Pipeline-0.11 from CPAN.
gitpan-cpan-distribution: Alvis-Pipeline
gitpan-cpan-version:      0.11
gitpan-cpan-path:         MIRK/Alvis-Pipeline-0.11.tar.gz
gitpan-cpan-author:       MIRK
gitpan-cpan-maturity:     released

+++ bin/alvis-zsink.pl
@@ -1,6 +1,6 @@
 #!/usr/bin/perl -w
 
-# $Id: alvis-zsink.pl,v 1.2 2006/08/14 18:00:32 mike Exp $
+# $Id: alvis-zsink.pl,v 1.6 2006/08/31 16:13:16 mike Exp $
 #
 # This is an Alvis Sink -- that is, a program that sits at the end of
 # an Alvis pipeline absorbing documents that are fed to it from an
@@ -12,12 +12,13 @@
 # The program needs to be told what port to listen on (for the
 # pipeline) and what host/port to write to (for the Z39.50 server).
 # For example, to listen on port 8021 and feed documents to a server
-# on localhost:9996, use:
-#	alvis-zsink.pl 8021 localhost:9996
+# on localhost:1314, use:
+#	alvis-zsink.pl 8021 localhost:1314
 
 use strict;
 use warnings;
 use Alvis::Pipeline 0.07;
+use Net::Z3950::ZOOM 1.11;	# We don't use this, but it's ZOOM's release
 use ZOOM;
 
 if (@ARGV != 2 && @ARGV != 4) {
@@ -25,7 +26,7 @@ if (@ARGV != 2 && @ARGV != 4) {
 Usage: $0 <listen-port> <z-host> [<user> <password>]
     <listen-port> is the port on which to listen for documents being
     fed down the Alvis pipeline; <z-host> is ZOOM-style Z39.50 host
-    string such as 'localhost:9996' or 'tcp:alvis.indexdata.com:8122'
+    string such as 'localhost:1314' or 'tcp:alvis.indexdata.com:8122'
 __EOT__
     exit 1;
 }
@@ -35,30 +36,43 @@ my $pipe = new Alvis::Pipeline::Read(port => $port,
 				     spooldir => "/tmp/alvis-spool")
     or die "can't create read-pipe on port $port: $!";
 $pipe->option(sleep => 1);
+print "Listening on pipeline\n";
 
 my $options = new ZOOM::Options();
 $options->option(user => $user) if defined $user;
 $options->option(password => $password) if defined $password;
 my $conn = create ZOOM::Connection($options);
-$conn->connect($zhost);
-print "connected: waiting for documents\n";
+my $connected = 0;
 
 my $n = 0;
 $| = 1;
 while (my $xml = $pipe->read(1)) {
+    eval {
+	$conn->connect($zhost);
+    }; if ($@ && ref $@ && $@->isa("ZOOM::Exception") &&
+	   $@->diagset() eq "Bib-1" && $@->code() == 224) {
+	# Do nothing: this allows for a bug in ZOOM-C (as of YAZ
+	# version 2.1.27) whereby a no-opping call to connect() does
+	# not clear any old error indication.
+	print "ignoring re-occurrence of old error in connect()\n";
+    } elsif ($@) {
+	die $@; # re-throw
+    }
+
+    if (!$connected) {
+	print "connected to Z39.50 server\n";
+	$connected = 1;
+    }
     print "got document ", ++$n, " ... ";
 
-  AGAIN:
     eval {
 	my $p = $conn->package();
 	$p->option(action => "specialUpdate");
 	$p->option(record => $xml);
-	# Could set "recordIdOpaque" if we had a record-ID
-	$p->send("update");
+	eval { $p->send("update"); };
+	if ($@) { die $@; }
 	print "sent package ... ";
 	$p->destroy();
-
-	sleep 1;
 	$p = $conn->package();
 	$p->option(action => "commit");
 	$p->send("commit");
@@ -66,7 +80,7 @@ while (my $xml = $pipe->read(1)) {
     };
 
     if (!$@) {
-	print STDERR "added document\n";
+	print "added document\n";
 	next;
     } elsif (!ref $@ || !$@->isa("ZOOM::Exception")) {
 	# A non-ZOOM error, which is totally unexepected.  Treat this
@@ -74,18 +88,12 @@ while (my $xml = $pipe->read(1)) {
 	# that the spooling child process is killed.
 	$pipe->close();
 	die $@;
-    } elsif ($@->diagset() ne "ZOOM" ||
-	     $@->code() != ZOOM::Error::CONNECTION_LOST) {
-	# A ZOOM error other than connection lost, e.g. BIB-1 224,
-	# "ES: immediate execution failed".  Most such cases need not
-	# be treated as fatal, so we just log it and continue.
-	warn "$@\n";
     } else {
-	# Connection lost, most likely because Zebra got bored and
-	# timed it out.  Re-forge the connection and try again.
-	warn "ZOOM connection lost (probably due to timeout): re-forging\n";
-	create ZOOM::Connection($options);
-	$conn->connect($zhost);
-	goto AGAIN;
+	# A ZOOM error, e.g. BIB-1 224, "ES: immediate execution
+	# failed".  Most such cases need not be treated as fatal, so
+	# we just log it and continue to listen for subsequent
+	# documents.
+	print "failed\n";
+	warn "discarding record: $@\n";
     }
 }
